==> capstone_babs/tripanimation/1-GenerateMilestonesInTopRoutes.ipynb <==
%matplotlib inline
import pandas as pd



import json

def generateMilestones(commute, milestoneLength):
    commuteMileStones = {}
    commuteMileStones['totalDistance'] = commute[0]['legs'][0]['distance']['value']
#     for step in commute[0]['legs'][0]['steps']:
#         print 'step', step['distance']['value']
    commuteMileStones['mileStoneLats'] = []
    commuteMileStones['mileStoneLngs'] = []

    totalSteps = commute[0]['legs'][0]['steps']
    
    currentStepIndex = 0
    currentStep = totalSteps[currentStepIndex]
    
    pendingDistance = commuteMileStones['totalDistance']
    pendingCurrentStepDistance = currentStep['distance']['value']
    pendingMileStoneDistance = milestoneLength
    
    currentStepCoveredDistance = 0

    milestoneLat = currentStep["start_location"]["lat"]
    milestoneLng = currentStep["start_location"]["lng"]
    commuteMileStones['mileStoneLats'].append(milestoneLat)
    commuteMileStones['mileStoneLngs'].append(milestoneLng)
            
    while (pendingDistance > 0):
#         print "Loop enter current step: ", currentStepIndex, ' pendingDistance: ', pendingDistance, ' pendingMileStoneDistance: ', pendingMileStoneDistance, ', currentStepCoveredDistance', currentStepCoveredDistance
        if (pendingMileStoneDistance >= pendingDistance):
#           Not much total distance left, wrap up and return
#             print 'final milestone'
            milestoneLat = currentStep["end_location"]["lat"]
            milestoneLng = currentStep["end_location"]["lng"]
            commuteMileStones['mileStoneLats'].append(milestoneLat)
            commuteMileStones['mileStoneLngs'].append(milestoneLng)
            return commuteMileStones
        if (pendingMileStoneDistance >= pendingCurrentStepDistance):
#             Not much distance left in current step, move over to next step with remaining milestone distance
            pendingMileStoneDistance = (pendingMileStoneDistance - pendingCurrentStepDistance)
            pendingDistance = (pendingDistance - pendingCurrentStepDistance)

            currentStepIndex = currentStepIndex + 1
            currentStep = totalSteps[currentStepIndex]
            pendingCurrentStepDistance = currentStep['distance']['value']
            currentStepCoveredDistance = 0
            
#             print 'move pver to the next step: ', currentStepIndex, ' pendingDistance: ', pendingDistance, ' pendingMileStoneDistance: ', pendingMileStoneDistance
            continue
          
        pendingCurrentStepDistance = pendingCurrentStepDistance - pendingMileStoneDistance
        pendingDistance = (pendingDistance - pendingMileStoneDistance)
        pendingMileStoneDistance = milestoneLength
        
        currentStepCoveredDistance = currentStepCoveredDistance + pendingMileStoneDistance
        currentStepLength = currentStep['distance']['value']

        startMilestoneLat = float(currentStep["start_location"]["lat"])
        startMilestoneLng = float(currentStep["start_location"]["lng"])

        endMilestoneLat = float(currentStep["end_location"]["lat"])
        endMilestoneLng = float(currentStep["end_location"]["lng"])
        
        incrementalLat = ((endMilestoneLat - startMilestoneLat)*currentStepCoveredDistance)/currentStepLength
        milestoneLat = str(startMilestoneLat + incrementalLat)
        
        incrementalLng = ((endMilestoneLng - startMilestoneLng)* currentStepCoveredDistance)/currentStepLength
        milestoneLng = str(startMilestoneLng + incrementalLng)
        
#         print 'Calculating Proportionate lat long ', 
#         print 'from',  startMilestoneLat, ':', startMilestoneLng
#         print 'to',  startMilestoneLat, ':', startMilestoneLng
#         print currentStepCoveredDistance, ' covered out of ', currentStepLength
#         print "incremental lat lngs: ", incrementalLat, ':', incrementalLng
#         print "Milestone lat lngs: ", milestoneLat, ':', milestoneLng
#         commuteMileStones['mileStones'].append((milestoneLat, milestoneLng))
        commuteMileStones['mileStoneLats'].append(milestoneLat)
        commuteMileStones['mileStoneLngs'].append( milestoneLng)
        
#         print 'current step: ', currentStepIndex, ', pendingDistance: ', pendingDistance, ', pendingMileStoneDistance: ', pendingMileStoneDistance
# startStationId, endStationId, distance, number of milestones, latitudes, longitudes

    return commuteMileStones



def getMileStoneRow(sampleCommuteJson, startStationId, endStationId):
    generatedCommuteMileStones = generateMilestones(sampleCommuteJson, 50)

    print 'start ',startStationId, 'end ', endStationId
    print 'totalDistance', generatedCommuteMileStones['totalDistance']    
    print 'Number of milestones', len(generatedCommuteMileStones['mileStoneLats'])

    # type(generatedCommuteMileStones['mileStones'])
#     print 'firsttotalDistance', sampleCommuteJson[0]['legs'][0]['distance']['value']
#     print 'Lat', sampleCommuteJson[0]['legs'][0]['start_location']['lat']

    generatedCommuteMileStones['startStationId']= startStationId
    generatedCommuteMileStones['endStationId']= endStationId
    generatedCommuteMileStones['milestoneCount'] = len(generatedCommuteMileStones['mileStoneLats'])
    return generatedCommuteMileStones


topRoutes = pd.DataFrame.from_csv('./data/2014_SF_Top29Routes.csv')

stations = pd.DataFrame.from_csv('./data/201402_station_data.csv', index_col='station_id', encoding='utf-8')

# print 'Top Routes:', len(topRoutes.index)

# print 'Station Count:', len(stations.index)
# stations['lat'][69]
import csv
import googlemaps
import datetime
from datetime import timedelta


gmaps = googlemaps.Client(key='AIzaSyC08BfStY-4KCNCe9rJS438VSvB5BC-wag')



# originStationLocation = {'lat': stations['lat'][69], 'lng':stations['long'][69]}
# destinationStationLocation = {'lat': stations['lat'][39], 'lng':stations['long'][39]}
    
# commute= gmaps.directions(originStationLocation, destinationStationLocation, 
#                         mode="bicycling",
#                           departure_time=datetime.datetime.now()  )

# #     print row['startStationId'], row['endStationId'], originStationLocation, destinationStationLocation, commute[0]['legs'][0]['distance']['value']
# #     print row['startStationId'], row['endStationId'], originStationLocation, destinationStationLocation

# milestoneRow = getMileStoneRow(commute, 69,39)
# milestoneRow


toCSV=[]
for index, row in topRoutes.iterrows():
    originStationLocation = {'lat': stations['lat'][row['startStationId']], 'lng':stations['long'][row['startStationId']]}
    destinationStationLocation = {'lat': stations['lat'][row['endStationId']], 'lng':stations['long'][row['endStationId']]}
    
    commute= gmaps.directions(originStationLocation, destinationStationLocation, 
                        mode="bicycling",
                          departure_time=datetime.datetime.now()  )
    print
    print row['startStationId'], row['endStationId'], originStationLocation, destinationStationLocation, commute[0]['legs'][0]['distance']['value']

    milestoneRow = getMileStoneRow(commute, row['startStationId'],row['endStationId'])
    toCSV.append(milestoneRow)
#     # keys = milestoneRow.keys()
    keys=[
     'startStationId',
     'endStationId',
     'totalDistance',
     'milestoneCount',
     'mileStoneLats',
     'mileStoneLngs']
    with open('./data/2014_SF_Top_TripMileStones.csv', 'wb') as output_file:
        dict_writer = csv.DictWriter(output_file, keys)
        dict_writer.writeheader()
        dict_writer.writerows(toCSV)
#     keys
