==> capstone_babs/tripanimation/4-PrepareGmapTimeLine.ipynb <==
import pandas as pd

# data=pd.DataFrame.from_csv('./data/SortedSampleTimeline.csv', dtype = {'lat': str})
data=pd.read_csv('./data/2014_SF_Aug_TopTimeline.csv', dtype = {'lat': str, 'lng': str})
preparedGmapTimeline = data.groupby(['time']).agg(lambda x: ','.join(x))

preparedGmapTimeline.to_csv('./data/2014_SF_Aug_GmapTimeline.csv')
# data.dtypes
# data
