==> capstone_babs/tripanimation/5-CreateGmaps.ipynb <==
import pandas as pd
import gmplot

gmapTimeline = pd.read_csv('./data/2014_SF_Aug_GmapTimeline.csv', dtype = {'lat': str, 'lng': str})

SFStationLats=37.783871,37.795001,37.79728,37.794231,37.795425,37.788975,37.799953,37.789625,37.795392,37.791464,37.787152,37.789756,37.792251,37.781752,37.77865,37.781332,37.80477,37.780526,37.785299,37.786978,37.782259,37.771058,37.774814,37.776619,37.784878,37.776377,37.776617,37.788446,37.780356,37.798522,37.794139,37.7913,37.786305,37.789625,37.798541
SFStationsLngs=-122.408433,-122.39997,-122.398436,-122.402923,-122.404767,-122.403452,-122.398525,-122.390264,-122.394203,-122.391034,-122.388013,-122.394643,-122.397086,-122.405127,-122.418235,-122.418603,-122.403234,-122.390288,-122.396236,-122.398108,-122.392738,-122.402717,-122.418954,-122.417385,-122.401014,-122.39607,-122.39526,-122.408499,-122.412919,-122.407245,-122.394434,-122.399051,-122.404966,-122.400811,-122.400862
    

gmap = gmplot.GoogleMapPlotter(37.785848,-122.401505,14)

for index, row in gmapTimeline.iterrows():
    
    lats = [float(x) for x in row['lat'].split(",")]
    lngs = [float(x) for x in row['lng'].split(",")]

    pathlat = tuple(lats)
    pathlon = tuple(lngs)
    
    gmap = gmplot.GoogleMapPlotter(37.785848,-122.401505,14)
    
    gmap.scatter(pathlat, pathlon, '#FF0000', size=25, marker=False)
    gmap.scatter(SFStationLats, SFStationsLngs, '#0000FF', size=25, marker=False)
    filename = row['time'].replace(" ", "_")
    filename = filename.replace("/", "-")
    filename = filename.replace(":", "-")
    filename = './2013_Aug_tripGmapSnapShots/' + filename + '.html'
    gmap.draw(filename)
print 'done'
